revert the copyrigh, only when commit feature, add the copyright field

+++ trunk/src/core/srs_core.cpp
@@ -43,7 +43,7 @@ void srs_update_system_time_ms()
     gettimeofday(&now, NULL);
 
     // we must convert the tv_sec/tv_usec to int64_t.
-    _srs_system_time_us_cache = now.tv_sec * 1000 * 1000 + now.tv_usec;
+    _srs_system_time_us_cache = ((int64_t)now.tv_sec) * 1000 * 1000 + (int64_t)now.tv_usec;
     
     _srs_system_time_us_cache = srs_max(0, _srs_system_time_us_cache);
 }
